cli: Split commands.rs by domain
